cloud_functions/etl_to_bq/requirements.txt
google-cloud-bigquery==1.19.0
pyyaml
googlemaps==4.2.0